Morph/requirements.txt
albumentations==0.4.3
decord==0.6.0
einops==0.8.0
iopath==0.1.10
matplotlib==3.7.1
numpy==1.24.3
omegaconf==2.3.0
opencv_python_headless==4.5.5.64
pandas==2.0.2
peft==0.2.0
Pillow==9.5.0
Pillow==10.3.0
pytorch_lightning==1.0.8
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
scipy==1.13.0
skimage==0.0
timm==0.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
webdataset==0.2.48
xformers==0.0.20
